Users can now define custom predicates to use in game descriptions.
Predicates are now functors declared in predicates.hpp and implemented
in predicates.cpp. The predicates are mapped to names using the
get_predicate map of the Predicates class. These names can be used in
game descriptions. For now, predicates have access to the current
Environment and next position. This is a step towards being able to
define custom predicates in game descriptions.