Users can now define custom side effects to use in game descriptions.
SideEffects are functors declared in side_effects.hpp and implemented in
side_effects.cpp. The side effects are mapped to names using the
get_side_effect map of the SideEffects class. These names can be used in
game descriptions. Side effects have access to the environment, the cell
a piece is leaving, and the cell a piece is going to. The side effect
describes what happens during the move. If no side effect is provided,
the default side effect is to move the piece and capture any piece on
the destination cell.