Users can now define post conditions that must hold after a player makes a move.
An example of such a post condition is "the king must not be attacked
after you move" in chess. Post conditions are defined in the game
description language using regular expressions, and the post conditions
holds if the regular expression is not matched. Post conditions are
defined for player/piece pairs. After a player makes a move, all its
post conditions are checked for each of the pieces they are defined for.