Move state evaluation function from Agent to Model and more reasonable hyperparameters.
Moving the responsibility of evaluating a state from the Agent to its
neural network allows other modules to evaluate a state without
knowledge of the agent. Useful for implementing MCTS.

The neural network does not need so many epochs to learn from training
data and searching too deep with negamax takes forever.